Remove HAVE_STRERROR, no longer used in VisIt. Remove various unreferenced variables and other compiler warnings.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@13991 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/IVP/MemStream.h
@@ -132,7 +132,7 @@ struct IVP_API MemStream
     friend std::ostream& operator<<(std::ostream &out, const MemStream &m)
     {
         out<<" MemStream(p= "<<m.pos()<<", l= "<<m.len()<<"["<<m.capacity()<<"]) data=[";
-        for (int i=0; i < m.len(); i++)
+        for (size_t i=0; i < m.len(); i++)
             out<<(int)(m._data[i])<<" ";
         out<<"]";
         return out;
