+++ src/common/expr/ExprNode.C
@@ -297,7 +297,7 @@ ListExpr::ExtractNumericElements(vector<double> &output)
 
     std::vector<ListElemExpr*> *elems = GetElems();
 
-    for (int i = 0 ; i < elems->size() ; i++)
+    for (size_t i = 0 ; i < elems->size() ; i++)
     {
         ExprNode *item = (*elems)[i]->GetItem();
         if (item->GetTypeName() == "FloatConst")
@@ -329,7 +329,7 @@ ListExpr::ExtractStringElements(vector<std::string> &output)
 
     std::vector<ListElemExpr*> *elems = GetElems();
 
-    for (int i = 0 ; i < elems->size() ; i++)
+    for (size_t i = 0 ; i < elems->size() ; i++)
     {
         ExprNode *item = (*elems)[i]->GetItem();
         if (item->GetTypeName() == "StringConst")
